requirements.txt
asgiref==3.4.1
dataclasses==0.6
Django==3.2.6
django-ckeditor==6.1.0
django-environ-2==2.1.0
django-js-asset==1.2.2
django-taggit==1.5.1
Pillow==8.3.1
python-dotenv
pytz==2021.1
sqlparse==0.4.2
djangorestframework==3.11.1
django-filter
djoser
drf-yasg